Move the `builtin` helper
to `shader/execution/expression/call/builtin/builtin.ts`

This is specific to builtins.
