+++ src/webgpu/shader/execution/expression/call/builtin/any.spec.ts
@@ -13,7 +13,9 @@ import {
   vec3,
   vec4,
 } from '../../../../../util/conversion.js';
-import { builtin, run } from '../../expression.js';
+import { run } from '../../expression.js';
+
+import { builtin } from './builtin.js';
 
 export const g = makeTestGroup(GPUTest);
