+++ src/webgpu/shader/execution/expression/call/builtin/atan.spec.ts
@@ -8,7 +8,9 @@ import { ulpThreshold } from '../../../../../util/compare.js';
 import { kBit, kValue } from '../../../../../util/constants.js';
 import { f32, f32Bits, TypeF32 } from '../../../../../util/conversion.js';
 import { biasedRange, linearRange } from '../../../../../util/math.js';
-import { builtin, Case, Config, run } from '../../expression.js';
+import { Case, Config, run } from '../../expression.js';
+
+import { builtin } from './builtin.js';
 
 export const g = makeTestGroup(GPUTest);
