+++ src/webgpu/shader/execution/expression/call/builtin/atan2.spec.ts
@@ -9,7 +9,9 @@ import { ulpThreshold } from '../../../../../util/compare.js';
 import { kValue } from '../../../../../util/constants.js';
 import { f32, TypeF32 } from '../../../../../util/conversion.js';
 import { biasedRange, linearRange } from '../../../../../util/math.js';
-import { builtin, Case, Config, run } from '../../expression.js';
+import { Case, Config, run } from '../../expression.js';
+
+import { builtin } from './builtin.js';
 
 export const g = makeTestGroup(GPUTest);
