+++ src/webgpu/shader/execution/expression/call/builtin/clamp.spec.ts
@@ -19,7 +19,9 @@ import {
   u32Bits,
 } from '../../../../../util/conversion.js';
 import { isSubnormalScalar } from '../../../../../util/math.js';
-import { builtin, Case, Config, run } from '../../expression.js';
+import { Case, Config, run } from '../../expression.js';
+
+import { builtin } from './builtin.js';
 
 export const g = makeTestGroup(GPUTest);
