+++ src/webgpu/shader/execution/expression/call/builtin/extractBits.spec.ts
@@ -15,7 +15,9 @@ import {
   vec4,
   TypeVec,
 } from '../../../../../util/conversion.js';
-import { builtin, Config, run } from '../../expression.js';
+import { Config, run } from '../../expression.js';
+
+import { builtin } from './builtin.js';
 
 export const g = makeTestGroup(GPUTest);
